requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
PyYAML==3.13
requests==2.21.0
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
sphinx==1.8.3
ed25519==1.4